LayoutTests/platform/ios-sim-deprecated/accessibility/tab-role-expected.txt

tab1

This test makes sure that role='tab' shows up as a distinct element from button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tab.iphoneLabel is 'AXLabel: tab1'
PASS button.iphoneTraits != tab.iphoneTraits is true
PASS tab.iphoneTraits != 0 is true
FAIL tab.isSelected should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
